+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [TBA] - ?
+
+## Changed
+
+* Added backup and restore of IMAP flags (e.g. "Seen", "Draft").
+* Introduced a new metadata format.
+* Included data about message offsets and lengths in the new metadata to
+  facilitate for future integrity checks.
+* Added a migrator to transform the old (version 2) metadata files
+  into the newer (version 3) files.
+
 ## [6.2.1] - 2022-07-12
 
 ## Changed